# src/openapi.yaml
openapi: 3.0.2
info:
  x-logo:
    href: https://kortxyz.github.io/favicon.ico
  title: OGCAPI
  version: 0.0.1
  description: Implemations of OGC API
  contact:
    name: Tino Kastbjerg Stigsen
    email: info@kort.xyz
    url: https://kort.xyz
  license:
    name: CC-BY 4.0 license
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - url: https://localhost:3000
...
paths:
...
  /collections/{collectionId}/styles:
    get:
      operationId: getCollectionStyles
      tags:
        - Styles
      summary: get Styles
      description: Fetches all styles for collection
      parameters:
        - $ref: "#/components/parameters/collectionId"
      responses:
        "200":
          description: Style
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /collections/{collectionId}/styles/{styleId}:
    get:
      operationId: getCollectionStyle
      tags:
        - Styles
      summary: get Style
      description: Fetches a style for collection
      parameters:
        - $ref: "#/components/parameters/collectionId"
        - $ref: "#/components/parameters/styleId"
      responses:
        "200":
          description: Style
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
...
  /styles:
    get:
      tags:
        - Styles
      summary: information about the available styles
      operationId: getStyles
      description: |-
        This operation fetches the set of styles available. For
        each style the id, a title, links to the stylesheet of
        the style in each supported encoding, and the link to the
        metadata is provided.

        Testbed-15 only requires support for a small number of the
        styles. Therefore, the currently simple approach is sufficient,
        but in general the operation should support paging (using a
        parameter `limit` and links to the `next` page in responses).
      parameters:
        - $ref: "#/components/parameters/fCommon"
      responses:
        200:
          description: the set of available styles
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                  styles:
                    type: array
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
    post:
      tags:
        - Styles
      summary: adds a new style
      operationId: postStyle
      description: |-
        Adds a style to the style repository or just validates a style.

        If the parameter validate is set to yes, the style will be validated before adding the style to the server. If the parameter validate is set to only, the server will not be changed and only the validation result will be returned.

        In case, a new style is created, the following rules apply.

        If the style submitted in the request body includes an identifier (this depends on the style encoding), that identifier will be used. If a style with that identifier already exists, an error is returned.

        If no identifier can be determined from the submitted style, the server will assign a new identifier to the style.

        A minimal style metadata resource is created at /styles/{styleId}/metadata. Please update the metadata using a PUT request to keep the style metadata consistent with the style definition.

        The URI of the new style is returned in the header Location.

        This operation is only available to registered style authors.
      parameters:
        - $ref: "#/components/parameters/validate"
      responses:
        "201":
          description: style created
          headers:
            Location:
              schema:
                type: string
              description: URI of the new style
        "204":
          description: style validated successfully, no style has been created
        "400":
          description: invalid input, the style is invalid
        "401":
          description: unauthorized access
        "409":
          description: a style with that id already exists
      requestBody:
        description: Stylesheet to be added
        content:
          application/vnd.mapbox.style+json:
            schema:
              $ref: "#/components/schemas/mbstyle"
      security:
        - {}
  /styles/{styleId}:
    get:
      tags:
        - Styles
      summary: fetch a style by id
      description: |-
        Fetches the style with identifier `styleId`. The set of
        available styles can be retrieved at `/styles`.

        Not all styles are available in all style encodings.
      operationId: getStyle
      parameters:
        - $ref: "#/components/parameters/styleId"
        - $ref: "#/components/parameters/fStyle"
      responses:
        200:
          description: The style
          content:
            text/html:
              schema:
                type: string
            application/vnd.mapbox.style+json:
              schema:
                $ref: "#/components/schemas/mbstyle"
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
    put:
      tags:
        - Manage styles
      summary: replace a style or add a new style
      description: |-
        Replace an existing style with the id `styleId`. If no
        such style exists, a new style with that id is added.

        If the parameter `validate` is set to `yes`, the
        style will be validated before adding/updating the
        style. If the parameter `validate` is set to
        `only`, the server will not be changed and only the
        validation result will be returned.

        For updated styles, the style metadata resource at
        `/styles/{styleId}/metadata` is not updated.
        For new styles a minimal style metadata resource is
        created, too. Please update the metadata using a PUT
        request to keep the style metadata consistent with
        the style definition.

        This operation is only available to registered style
        authors.
      operationId: updateStyle
      security:
        - BasicAuth: []
      parameters:
        - $ref: "#/components/parameters/styleId"
        - $ref: "#/components/parameters/validate"
      requestBody:
        description: A single style in one of the supported style encodings.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/mbstyle"
      responses:
        "204":
          description: style updated, created or validated successfully
        "400":
          description: invalid input, the style is invalid
        "401":
          description: unauthorized access
  /resources:
    get:
      tags:
        - Resources
      summary: information about the available resources (symbols, sprites)
      operationId: getResources
      parameters:
        - $ref: "#/components/parameters/fCommon"
      responses:
        200:
          description: the set of available resources
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
  /resources/{resourceId}:
    get:
      tags:
        - Resources
      summary: fetch a symbol resource by id
      operationId: getResource
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        200:
          description: The symbol resource
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
    post:
      tags:
        - Resources
      summary: create or replace a resource by id
      operationId: postResource
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
      responses:
        201:
          description: Resource created
          headers:
            Location:
              schema:
                type: string
        400:
          $ref: "#/components/responses/ServerError"
      security:
        - {}
components:
  parameters:
    fCommon:
      name: f
      in: query
      description: >-
        Select the output format of the response. If no value is provided, the
        standard HTTP rules apply, i.e., the accept header will be used to
        determine the format.
      required: false
      schema:
        type: string
        enum:
          - json
          - html
...
    fStyle:
      name: f
      in: query
      description: >-
        Select the output format of the response. If no value is provided, the
        standard HTTP rules apply, i.e., the accept header will be used to
        determine the format.
      required: false
      schema:
        type: string
        enum:
          - mbs
          - html
...
    collectionId:
      name: collectionId
      in: path
      description: local identifier of a collection
      required: true
      schema:
        type: string
...
    styleId:
      name: styleId
      in: path
      description: Id of the style
      required: true
      schema:
        type: string
    resourceId:
      name: resourceId
      in: path
      description: Id of the resource
      required: true
      schema:
        type: string
    validate:
      name: validate
      in: query
      description: |-
        (part of conformance class 'style-validation')
        'yes' creates a new style after successful validation
        and returns 400, if validation fails,
        'no' creates the style without validation and
        'only' just validates the style without creating a
        new style and returns 400, if validation fails,
        otherwise 204.
      required: false
      style: form
      explode: false
      schema:
        type: string
        enum:
          - "yes"
          - "no"
          - only
      example: only
  schemas:
...
    exception:
      type: object
      description: >-
        Information about the exception: an error code plus an optional
        description.
      required:
        - statusCode
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
      example:
        statusCode: 404
        error: Not Found
        message: Route GET:/tileMatrixSets not found
...
    mbstyle:
      required:
        - layers
        - sources
        - version
      type: object
      properties:
        version:
          type: number
          example: 8
        name:
          type: string
          example: night
        sources:
          type: object
          properties:
            daraa:
              type: object
              properties:
                type:
                  type: string
                  example: vector
                url:
                  type: string
                  example: >-
                    https://services.interactive-instruments.de/vtp/daraa/tiles/default/{z}/{y}/{x}?f=mvt
        sprite:
          type: string
          example: >-
            http://vtp2018.s3-eu-west-1.amazonaws.com/static/mapstorestyle/sprites/sprites
        glyphs:
          type: string
          example: http://fonts.openmaptiles.org/{fontstack}/{range}.pbf
        layers:
          type: array
          items:
            type: object
...
  responses:
...
    ServerError:
      description: A server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/exception"
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic